+++ src/Facebook/Exceptions/FacebookResumableUploadException.php
@@ -30,4 +30,39 @@ namespace Facebook\Exceptions;
  */
 class FacebookResumableUploadException extends FacebookSDKException
 {
+    protected $startOffset;
+
+    protected $endOffset;
+
+    /**
+     * @return int|null
+     */
+    public function getStartOffset()
+    {
+        return $this->startOffset;
+    }
+
+    /**
+     * @param int|null $startOffset
+     */
+    public function setStartOffset($startOffset)
+    {
+        $this->startOffset = $startOffset;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getEndOffset()
+    {
+        return $this->endOffset;
+    }
+
+    /**
+     * @param int|null $endOffset
+     */
+    public function setEndOffset($endOffset)
+    {
+        $this->endOffset = $endOffset;
+    }
 }